requirements.txt
contourpy==1.1.1
cycler==0.12.1
DateTime==5.5
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1719234999090/work
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020413938/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
packaging==24.1
pandas==2.0.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1630696607296/work
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sklearn-model==0.0.6
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
zope.interface==7.0.1
